refact(ALL): FIX plugin-hashes with digest-factories
MAJOR REVAMP revamp to solve the bug discovered in a21bf38
about reusing plugin-digesters.

In the previous version, each "digester" was a 2-tuple
`(digester-instance, final-hash-func)`.
To solve the reuse bug without refactoring it would require to
re-initialize the plugins for each input.  

In this revision, each registered *digester* is actually a
`factory_function(fsize: int)` that will create a *digester* class with
2 methods:
- update(bytes)
- hexdigest() -> str # lower

NOTE that the factory-function takes `fsize` as its argument, - this is
necessary for git-digesters not to always slurp bytes, particularly for
URL-resources; so all other digesters must use a "special" factory that
ignores the `fsize` arg.


Other changes, impossible to separate in commits:

- FIX plugins - multiple problems were preventing them from running -
added travis TCs to detect them.
- Add `-x family` option to exclude families.
- The inclusion/exclusion logic is implemented within a class.
- Do not git-slurp if URL provide Content-Length.
- Avoid needless instantiation of excluded digesters.
- Avoid some top-level imports, to speed up cmd-line launch for
help-msg.